# frontend/config/swagger.yaml
swagger: '2.0'
info:
  description: 'This is the theme parks API spec.'
  version: '1.0.0'
  title: 'Themepark API'
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'com.themeparks.api'
...
schemes:
- 'https'
- 'http'
paths:
...
  /reservations:
    get:
      tags:
        - reservation
      operationId: getReservations
      parameters:
      - name: userId
        in: query
        description: user account
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Reservations'
    post:
      tags:
        - reservation
      description: create new reservations
      operationId: createReservation
      parameters:
      - name: body
        in: body
        description: reservation
        schema:
          $ref: '#/definitions/Reservations'
        required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Reservations'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      x-codegen-request-body-name: body
    delete:
      tags:
        - reservation
      description: delete reservation
      operationId: deleteReservation
      parameters:
        - name: reservationId
          in: query
          description: id of reservation to delete
          type: string
          required: true
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Reservations'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /reservations/confirm:
    post:
      tags:
        - reservation
      description: confirm reservation
      operationId: confirmReservation
      parameters:
      - name: reservationId
        in: query
        description: id of reservation to confirm
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Reservation'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  /schedules:
    get:
      tags:
        - schedule
      operationId: getSchedules
      parameters:
      - name: attractionId
        in: query
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Schedules'
    post:
      tags:
        - schedule
      description: create a new schedule
      operationId: createSchedule
      parameters:
      - in: 'body'
        name: 'body'
        required: true
        schema:
          $ref: '#/definitions/Schedule'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Schedules'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      tags:
        - schedule
      description: update a schedule
      operationId: updateSchedule
      parameters:
      - in: 'body'
        name: 'body'
        required: true
        schema:
          $ref: '#/definitions/Schedule'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Schedules'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - schedule
      description: delete a schedule
      operationId: deleteSchedule
      parameters:
        - name: scheduleId
          in: query
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Schedules'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /schedules/upcomingSchedule:
    get:
      tags:
        - schedule
      operationId: getUpcomingSchedule
      parameters:
      - name: attractionId
        in: query
        type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UpcomingSchedule'
    put:
      tags:
        - schedule
      operationId: updateUpcomingSchedule
      parameters:
      - in: 'body'
        name: 'body'
        required: true
        schema:
          $ref: '#/definitions/UpcomingSchedule'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UpcomingSchedule'
...
definitions:
...
  Reservations:
    type: array
    items:
      $ref: '#/definitions/Reservation'
  Reservation:
    type: object
    properties:
      id:
        type: string
      attraction:
        type: string
      user:
        type: string
      schedule:
        type: string
      startTime:
        type: string
        format: date-time
      endTime:
        type: string
        format: date-time
      location:
        type: string
      status:
        type: string
        enum: [waiting, passed, cancelled, confirmed, checking-in]
...
  ErrorResponse:
    required:
      - errorCode
    type: object
    properties:
      errorCode:
        type: string
        description: error code
      errorMessage:
        type: string
        description: error message
      success:
        type: boolean
        description: success or not
  Schedule:
    type: object
    properties:
      id:
        type: string
      attraction:
        type: string
      totalSeats:
        type: integer
      availableSeats:
        type: integer
      startTime:
        type: string
        format: date-time
      endTime:
        type: string
        format: date-time
      status:
        type: string
        enum: [upcoming, checking-in, post-checkin, completed, cancelled]
  Schedules:
    type: array
    items:
      $ref: '#/definitions/Schedule'
  UpcomingSchedule:
    type: object
    properties:
      id:
        type: string
      attraction:
        type: string
      totalSeats:
        type: integer
      availableSeats:
        type: integer
      startTime:
        type: string
        format: date-time
      endTime:
        type: string
        format: date-time
      nextNumberToCall:
        type: integer
      status:
        type: string
        enum: [upcoming, checking-in, post-checkin, completed, cancelled]
...